Implement basic parsing as proof of concept
Note: This is just a proof of concept, refactoring needed. In
particular, the parsing should be modified in such a way that the input
into function is a slice (vector) reference containing lines of string.
This is needed in order to support multi-line headings.

Also, multi-line headings support must be implemented.