Restructure the project according to design choice
Project is restructured as discussed in Development team discussion
board (at least partially). The HeadingBlock is moved into
'um_elements', traits for frontend are created.

Keep in mind, this is the beginning stage and many parts may change in
the future, but this should serve as a pattern on how to implement the
parsing of the Unimarkup syntax.