Avoid using lifetimes by simplifying UmSyntaxError
Note: This performs a String allocation, but by design program should
break as soon as we encounter an error. Therefore, this allocation
should not be problematic.

Note #2: Lifetimes and generally more borrowing might be introduced at a
later point in time, as this should perform better. Right now, focus is
on creating code which works, and since the unimarkup-rs should run only
once and create the output we are not concerned with String allocations
right now.