Add function for counting symbols to parser.rs
Function `count_symbol_until` is added to `parser.rs`.
This function counts the repetitions of wanted symbol from beginning of
the line, and checks if the symbol directly after the sequence matches
the given constrained.
The constraints are checked by providing a function as a parameter to
this function. The provided function performs wanted checks, i.e. one
such function is `char::is_whitespace(self: char)` which takes a
character and checks if it's whitespace, as defined by unicode spec.