fix: remove out-file path validation
Previous output file path validation did not allow writing to a new file
which doesn't exist yet. It only allowed writing to existing files. This
check is not needed, since the OS returns an error anyway if the given
output path is not valid, e.g. leading directories do not exist.