feat: cache token when split during parsing
During parsing it may occur that the closing token is ambiguous one, but
it's possible to resolve it. In such case, one part is used as the
closing token, but the next part should be considered as (probably)
opening one in next iteration. One such example:

**This is bold***And this is italic*

In that case, the ambiguous token (***) is split into two parts. First
the bold part is parsed, and then the italic part. This produces
following inlines:

**This is bold**|*And this is italic*
^^^^^^^^^^^^^^^^ ^^^^^^^^^^^^^^^^^^^^
      BOLD              ITALIC

After parsing the bold inline, ambiguous token is split and the second
part (in this case one star *) must be observed in next iteration ->
during parsing of next inline. It's therefore necessary to cache it, and
use the cached token first when parsing next inline.