feat: add inline caching and merging to parser
Add merging capability to the inline parser. This is needed in cases
where one inline is parsed until the next one, where they don't have the
same kind. After fully parsing the next inline it gets converted into
some other inline kind that may be the same as the previous one. This
produces two distinct inlines that could have been merged.
Since the parser is implemented as an `Iterator`, it is necessary to
parse the next inline before deciding whether the two can be merged.
Merging is not always possible, and sometimes only part of the inline is
merged. That is why the caching is also needed.