fix: handle case where tokens remain after parsing nested inline
It might be the case that some nested inline is detected and parser
starts the parsing. However, the closing token is never encountered.
This effectively leaves some tokens in parser stack, and they must be
parsed as a `Inline::Plain`. The rest of the parsed inlines (if any) are
stored into inline cache to be returned in next iteration.