perf: use `VecDeque` for `NestedContent`
Use `VecDeque` for `NestedContent` instead of a `Vec`. The `Vec` is
generally ok, but in this case insertion at the front of the collection
is sometimes necessary. The `Vec::insert` function can be used, but its
time complexity is `O(n)` where `n` is the number of elements after the
index used in `Vec::insert`. Since we insert at the front of the `Vec`,
that means that all elements must be shifted and that's not good.

`VecDeque::push_front` is the alternative, and it runs in `O(1)` so it
should ever so slighlty optimize the code.