feat: use default locales data and fallback locales
Use default locales data with english, austrian and bosnian languages.
Also, download fallback locales for provided locales. If locale provided
uses region extension that is invalid or default one, the data won't be
downloaded correctly. For example 'de-DE' or 'bs-BA' won't be
downloaded, and data would be missing keys. We extend the locales
provided with fallback locales consisting only of language without
extensions. For example fallback for 'de-DE' is 'de', and for 'bs-BA' is
'bs'. This makes sure we have at least the language data necessary for
our use case.