feat: use custom test harness to report which snapshot test failed
Previously, if a snap test failed in some other way rather than not
matching the snapshot file, we did not get a report on which test
failed. This is because we generate test cases dynamically (read files
and parse tests from them etc.). Using custom harness, we can also
generate output for each test case.

Note: cargo-nextest is recommended, since it reports these tests in a
nice way.